+++ srai/embedders/highway2vec/embedder.py
@@ -4,7 +4,7 @@ Highway2Vec embedder.
 This module contains the embedder from the `highway2vec` paper [1].
 
 References:
-    [1] https://doi.org/10.1145/3557918.3565865
+    1. https://doi.org/10.1145/3557918.3565865
 """
 from typing import Any, Dict, Optional