+++ srai/joiners/intersection_joiner.py
@@ -20,7 +20,7 @@ class IntersectionJoiner(Joiner):
     """
 
     def transform(
-        self, regions: gpd.GeoDataFrame, features: gpd.GeoDataFrame, return_geom: bool = True
+        self, regions: gpd.GeoDataFrame, features: gpd.GeoDataFrame, return_geom: bool = False
     ) -> gpd.GeoDataFrame:
         """
         Join features to regions based on an 'intersects' predicate.
@@ -30,7 +30,8 @@ class IntersectionJoiner(Joiner):
         Args:
             regions (gpd.GeoDataFrame): regions with which features are joined
             features (gpd.GeoDataFrame): features to be joined
-            return_geom (bool): whether to return geometry of the joined features
+            return_geom (bool): whether to return geometry of the joined features.
+                Defaults to False.
 
         Returns:
             GeoDataFrame with an intersection of regions and features, which contains