+++ srai/loaders/osm_loaders/__init__.py
@@ -1,6 +1,7 @@
 """OSM Loaders."""
 
+from ._base import OSMLoader
 from .osm_online_loader import OSMOnlineLoader
 from .osm_pbf_loader import OSMPbfLoader
 
-__all__ = ["OSMOnlineLoader", "OSMPbfLoader"]
+__all__ = ["OSMLoader", "OSMOnlineLoader", "OSMPbfLoader"]