+++ srai/utils/__init__.py
@@ -1,6 +1,21 @@
-"""Utilities."""
-# flake8: noqa F401
-from .download import *
-from .geocode import *
-from .geometry import *
-from .merge import *
+"""
+Module containing different utility functions.
+
+Those are either used internally by other modules, or can be used to simplify spatial data
+processing.
+"""
+from .download import download_file
+from .geocode import geocode_to_region_gdf
+from .geometry import buffer_geometry, flatten_geometry, flatten_geometry_series, remove_interiors
+from .merge import merge_disjointed_gdf_geometries, merge_disjointed_polygons
+
+__all__ = [
+    "download_file",
+    "geocode_to_region_gdf",
+    "buffer_geometry",
+    "flatten_geometry",
+    "flatten_geometry_series",
+    "remove_interiors",
+    "merge_disjointed_polygons",
+    "merge_disjointed_gdf_geometries",
+]